backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pandas
python-multipart
numpy
scikit-learn
requests
python-dotenv